third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-zIndex-non-auto-expected.txt

This is a testharness.js-based test.
FAIL z-index on non-positioned div assert_equals: Computed style expected "20" but got "auto"
PASS z-index on positioned div
Harness: the test ran to completion.
